Add: config setting

+++ src/main/java/portfolio/api/PortfolioRepository.java
@@ -4,5 +4,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Configuration
-public interface PortfolioRepository extends JpaRepository<Portfolio, Long> {
+public interface PortfolioRepository extends JpaRepository<Portfolio, Integer> {
+
 }
